fix: zero warnings build — all compiler warnings resolved
- Signed/unsigned comparisons in order_book_model, trade_tape_model
- Narrowing qint64/int in spread_chart std::max
- Removed dead emit_metrics_snapshot stub

+++ src/models/order_book_model.cpp
@@ -16,7 +16,7 @@ int OrderBookModel::columnCount(const QModelIndex&) const {
 }
 
 QVariant OrderBookModel::data(const QModelIndex& index, int role) const {
-    if (!index.isValid() || index.row() >= levels_.size()) return {};
+    if (!index.isValid() || index.row() >= static_cast<int>(levels_.size())) return {};
 
     const auto& entry = levels_[index.row()];
